Rename oauth2_provider to edx_oauth2_provider.
NOTE: Upgrading will require faking the initial migration.

django_oauth_toolkit uses the package name oauth2_provider, so we are
renaming our internal app to avoid naming conflicts

* Rename all imports
* Keep the table name the same.  It does not conflict with any table
  names from oauth2_provider.
* Update the initial migration to explicitly name the table.
* Bump version to 1.0.0

MA-1998: Install and configure django_oauth_toolkit to coexist with
existing oauth libraries.
