+++ edx_oauth2_provider/constants.py
