+++ edx_oauth2_provider/oidc/__init__.py
@@ -1,0 +1,6 @@
+"""
+OpenID Connect utilities.
+
+"""
+
+from .core import IDToken, id_token, userinfo
