+++ edx_oauth2_provider/tests/__init__.py
@@ -1,0 +1,5 @@
+# pylint: disable=missing-docstring
+
+from .base import OAuth2TestCase, IDTokenTestCase
+from .test_userinfo import UserInfoTestCase
+from .handlers import TestHandler
