+++ edx_oauth2_provider/tests/test_userinfo.py
@@ -1,5 +1,5 @@
 # pylint: disable=missing-docstring
-from oauth2_provider.tests.base import UserInfoTestCase
+from .base import UserInfoTestCase
 
 
 class UserInfoViewTest(UserInfoTestCase):
