+++ edx_oauth2_provider/utils.py
