--- oauth2_provider/oidc/__init__.py
@@ -1,6 +1,0 @@
-"""
-OpenID Connect utilities.
-
-"""
-
-from oauth2_provider.oidc.core import IDToken, id_token, userinfo
