--- oauth2_provider/tests/__init__.py
@@ -1,5 +1,0 @@
-# pylint: disable=missing-docstring
-
-from oauth2_provider.tests.base import OAuth2TestCase, IDTokenTestCase
-from oauth2_provider.tests.test_userinfo import UserInfoTestCase
-from oauth2_provider.tests.handlers import TestHandler
