+++ tests/urls.py
@@ -6,7 +6,7 @@ admin.autodiscover()
 urlpatterns = patterns(
     '',
     url(r'^admin/', include(admin.site.urls)),
-    url(r'^oauth2/', include('oauth2_provider.urls', namespace='oauth2')),
+    url(r'^oauth2/', include('edx_oauth2_provider.urls', namespace='oauth2')),
     url(r'^accounts/login/$', 'django.contrib.auth.views.login',
         {'template_name': 'login.html'}),
     url(r'^accounts/logout/$', 'django.contrib.auth.views.logout',
